+++ README.md
@@ -30,17 +30,15 @@ Have administrative privileges and run from cmd
 ```sh
 python -m venv --clear --prompt cs340 venv
 venv\Scripts\activate.bat
-pip install --upgrade pip setuptools wheel
-pip install --editable .
+pip install -e .
 ```
 
 ## Bash
 
 ```sh
-python3.10 -m venv --clear --prompt cs340 venv
+python3.11 -m venv --clear --prompt cs340 venv
 . venv/bin/activate
-pip install --upgrade pip setuptools wheel
-SETUPTOOLS_ENABLE_FEATURES="legacy-editable" pip install --editable .
+pip install -e .
 ```
 
 # Formatter
