--- setup.cfg
@@ -1,6 +1,0 @@
-[flake8]
-# ignore blank lines in empty inits
-per-file-ignores = __init__.py:W391
-
-max-complexity = 10
-max-line-length = 120
